+++ src/main/java/com/example/shoppingmall/order/dto/nonMemberOrderCheckDTO.java
@@ -1,0 +1,16 @@
+package com.example.shoppingmall.order.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class nonMemberOrderCheckDTO {
+
+    private Long checkNonMemberOrderNo;
+    private String checkNonMemberOrderName;
+
+}
